+++ src/app/layout.tsx
@@ -38,7 +38,7 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      <body className={`${inter.variable} ${ibmPlexMono.variable} font-sans bg-[#fafafa] text-black antialiased`}>
+      <body className={`${inter.variable} ${ibmPlexMono.variable} font-sans bg-galaxy-900 text-white antialiased`}>
         {children}
       </body>
     </html>